Fixed post_build script annoyances.

+++ external/post_build.sh
@@ -5,7 +5,7 @@ ROOTFS=$1
 cd ${ROOTFS}
 
 # Remove readline "old" files
-rm `find usr/lib -name '*.old'`
+find usr/lib -name '*.old' | xargs -r rm
 
 rm_links='lib32 usr/lib32'
 # Remove unneeded links.
@@ -22,28 +22,21 @@ done
 # Conditional links
 fake_files='etc/mdc_start.sh etc/run_script etc/run_script_ct
 	bin/thermal usr/sbin/hotMonitorStartScript.sh usr/sbin/hot_monitor'
-
 for link in ${fake_files}; do
    [[ ! -e $link ]] && ln -sr etc/log_attempt.sh ${link}
 done
 
 # Remove ld config files
-rm etc/ld.so.conf
-
-# Remove modprobe files but keep modules
-# rm `find lib/modules -name "modules.*"`
-
-# Remove xfs executables only want fsck.xfs and mkfs.xfs
-rm `find ./ -name "xfs_*"`
+[[ -f etc/ld.so.conf ]] && rm etc/ld.so.conf
 
 # Clearout var and tmp directories.
-cd var && rm -rf . && cd ..
-cd tmp && rm -rf . && cd ..
+cd var && rm -rf * && cd ..
+cd tmp && rm -rf * && cd ..
 
 rc_deps=${ROOTFS}/etc/init.d/rc.deps
 # Setup rc.deps and merge any rc.deps that exist.
 [ ! -f ${rc_deps} ] && mv ${rc_deps}.base ${rc_deps} || rm ${rc_deps}.base
-for dep_file in ${rc_deps}.*; do
+for dep_file in $(find `dirname ${rc_deps}` -name `basename ${rc_deps}`'.*'); do
 	cat ${dep_file} >>${rc_deps}
 	rm ${dep_file}
 done
